Sql: fix typo in the css table.
Ignore-this: a5740a1fe04656e177d9c15bda31c8c5

It was impossible to add more than one css to a wikipage. It was
impossible to add a css to a to a wikipage named X if a wikipage named
X in another wiki also have a css.

darcs-hash:20110811184743-da4f2-57dcfede986878a2311d238f1a2e8744f09e6cb5